+++ legacy_packages/unity-js-bridge/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @thirdweb-dev/unity-js-bridge
 
+## 0.6.17
+
+### Patch Changes
+
+- Updated dependencies [[`45677ad`](https://github.com/thirdweb-dev/js/commit/45677ad62012c63e3443f2e236311eb9a2bad094)]:
+  - @thirdweb-dev/chains@0.1.99
+  - @thirdweb-dev/sdk@4.0.68
+  - @thirdweb-dev/wallets@2.5.8
+  - @thirdweb-dev/auth@4.1.66
+
 ## 0.6.16
 
 ### Patch Changes